Integrate static URL tests into config_static_serving test
Move relative static URL tests from separate files into the existing
config_static_serving test suite for better organization.
