+++ e2e_playwright/config_static_serving.py
@@ -26,3 +26,31 @@ Images served via static serving:
 ![Streamlit](./app/static/streamlit-logo.png)
 """
 )
+
+# Test relative static URLs (/app/static/) in media elements
+# See: https://github.com/streamlit/streamlit/issues/12104
+
+st.header("Relative static URLs for media elements")
+
+# Test st.image with static URL
+st.subheader("st.image with /app/static/ URL")
+st.image("/app/static/streamlit-logo.png", caption="Streamlit logo from static")
+
+# Test st.audio with static URL
+st.subheader("st.audio with /app/static/ URL")
+st.audio("/app/static/cat-purr.mp3", format="audio/mp3")
+
+# Test st.video with static URL
+st.subheader("st.video with /app/static/ URL")
+st.video("/app/static/sintel-short.webm", format="video/webm")
+
+# Test st.chat_message with avatar using static URL
+st.subheader("st.chat_message with avatar from /app/static/")
+with st.chat_message("user", avatar="/app/static/streamlit-mark.png"):
+    st.write("Hello! My avatar comes from a static URL.")
+
+# Test st.logo with static URL
+st.subheader("st.logo with /app/static/ URL")
+st.logo("/app/static/streamlit-logo-small.png")
+
+st.success("All static URL elements rendered successfully!")
